Templates result in a seperate templates.xml
ERM46511
